moves models into organization folder
