+++ models/models.go
@@ -5,8 +5,7 @@ const (
 	WorkspaceUserCollectionName     = "workspace_users"
 	PluginCollectionName            = "plugins"
 	PluginCollectionsCollectionName = "plugin_collections"
-	OrganizationCollectionName      = "organizations"
-	MarkeplaceCollectionName        = "marketplaces" // ??
+	MarketplaceCollectionName        = "marketplaces" // ??
 	MessagesCollectionName          = "messages"
 	InstalledPluginsCollectionName  = "installed_plugins"
 )
